refactor: tokenRepository에서만 사용되는 UserDataDTO의 이름 및 위치 변경
